Moved redis_utils.py to FeatureToggle
